hype_pipeline/services/hourly_trade_data/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests>=2.32.0
urllib3==2.5.0
websocket-client>=1.8.0
